Added Drugs & Wires
Pretty sure it's correct, but do test! 

The "random" link is generated server-side - each page is served with a link to another comic (chosen at random) hardcoded into the href. I presume this is quite common, and PageFlip pulls a fresh link out of the DOM each time, in which case no problem :). If you tried to cache/memoize it, it'd always link to the same comic.